requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
geographiclib==2.0
geopy==2.4.1
google==3.0.0
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
lxml==5.1.0
numpy==1.26.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pymongo==3.12.3
pyparsing==3.1.2
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.2.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0